Refactor: Changed data type of "grandTotal" and return value of "GetTotalPrice()" to decimal, as this is a more suitable data type for storing amounts of money.

+++ Checkout.Lib/Checkout.cs
@@ -2,14 +2,14 @@
 {
     public class Checkout
     {
-        private int grandTotal = 0;
+        private decimal grandTotal = 0;
 
         public void Scan(string sku)
         {
             this.grandTotal = 40;
         }
 
-        public int GetTotalPrice()
+        public decimal GetTotalPrice()
         {
             return this.grandTotal;
         }
